Remove test_workflow_builder_func_env_var.py and fix tests
